+++ commands/__init__.py
@@ -1,0 +1,6 @@
+# Add here every command's filenames 
+__all__ = ["start", "help", "trans", "owo", "zalgos", "doggo", "inspire", "meme","audio", "spectral", "unsplash", "christmas", "pap", "slap", "aouguri"]
+
+from commands import *
+
+
