+++ commands/trans.py
@@ -1,0 +1,6 @@
+def init(update, context):
+    #log(update, lang["trans.log"])
+
+    # Send message
+    context.bot.send_message(
+        chat_id=update.effective_chat.id, text="https://bit.ly/34D6YDC")
