+++ errors/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["callback_error", "log"]
+
+from errors import *
